auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
email-validator
passlib[argon2]
argon2-cffi
python-dotenv
cryptography
